+++ modules/react/src/index.ts
@@ -9,6 +9,14 @@ export {default} from './deckgl';
 export {CompassWidget} from './widgets/compass-widget';
 export {FullscreenWidget} from './widgets/fullscreen-widget';
 export {ZoomWidget} from './widgets/zoom-widget';
+export {GeolocateWidget as _GeolocateWidget} from './widgets/geolocate-widget';
+export {InfoWidget as _InfoWidget} from './widgets/info-widget';
+export {LoadingWidget as _LoadingWidget} from './widgets/loading-widget';
+export {ResetViewWidget as _ResetViewWidget} from './widgets/reset-view-widget';
+export {ScaleWidget as _ScaleWidget} from './widgets/scale-widget';
+export {ScreenshotWidget as _ScreenshotWidget} from './widgets/screenshot-widget';
+export {SplitterWidget as _SplitterWidget} from './widgets/splitter-widget';
+export {ThemeWidget as _ThemeWidget} from './widgets/theme-widget';
 export {useWidget} from './utils/use-widget';
 
 // Types