+++ docs/whats-new.md
@@ -6,9 +6,13 @@ This page contains highlights of each deck.gl release. Also check our [vis.gl bl
 
 Target release date: Q2, 2025
 
+### Widgets
+
+- React: Pre-wrapped React components for the new deck.gl widgets are available via the [`@deck.gl/react`](./api-reference/react/overview.md) package, including: `ResetViewWidget`, `ScaleWidget`, `GeolocateWidget`, `ScreenshotWidget`, `LoadingWidget`, `ThemeWidget`, `InfoWidget`, and `SplitterWidget` (in addition to the `ZoomWidget`, `CompassWidget`, `FullscreenWidget` from v9.1)
+
 ## Core
 
-- [`View.clone()`](./api-reference/core/view.md) - New method that simplifies creating new Views with modified props.
+- [`View.clone()`](./api-reference/core/view.md) - New method that simplifies creating new Views with modified props, similar to `Layer.clone()`.
 
 ## deck.gl v9.1